add support for rename refactoring
For this I also added the FlatbuffersNamedElement interface
to a DeclaredType so that it could be renamed as well

This was trickier than I thought. In order to not crash when
asking a FlatbuffersNamedElement for it's ident, I needed to
actually replace the FlatbuffersIdent with IDENTIFIER in
flatbuffers.bnf. I presume this is the case because when
renaming an element, it briefly exists as a vanilla
identifier before being restructured by the parser into a
FlatbuffersIdent holding an identifier.